\name{GSEABase-package}
\alias{GSEABase-package}
\alias{GSEABase}
\docType{package}
\title{
  Gene set enrichment data structures and methods
}
\description{}
\details{}
\author{
  Written by Martin Morgan, Seth Falcon, Robert Gentleman. 
  Maintainer: Biocore Team <biocpkgs@lists.fhcrc.org>
}
\seealso{
  \linkS4class{GeneSet},
  \linkS4class{GeneColorSet}
}
\examples{
example(GeneSet)
}
\keyword{package}
